LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-dynamic-002-expected.txt


PASS .target 1
FAIL .target 2 assert_equals:
<div class="target" style="width: anchor-size(--a1 width)" data-expected-width="5"></div>
width expected 5 but got 0
PASS .target 3
FAIL .target 4 assert_equals:
<div class="target" style="width: anchor-size(--a2 width)" data-expected-width="5"></div>
width expected 5 but got 0
FAIL .after .target 5 assert_equals:
<div class="target" style="left: anchor(--a1 right)" data-offset-x="10"></div>
offsetLeft expected 10 but got 5
FAIL .after .target 6 assert_equals:
<div class="target" style="width: anchor-size(--a1 width)" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .after .target 7 assert_equals:
<div class="target" style="left: anchor(--a2 right)" data-offset-x="10"></div>
offsetLeft expected 10 but got 5
FAIL .after .target 8 assert_equals:
<div class="target" style="width: anchor-size(--a2 width)" data-expected-width="10"></div>
width expected 10 but got 0
